Update aura to turn edges
